using minilp instead of highs by default so install is easier. changed some default params, options, output logs. pre-merge push
