+++ src/graph_processing.rs
@@ -3,10 +3,9 @@ use crate::file_reader;
 use crate::global_clustering;
 use crate::local_clustering;
 use crate::part_block_manip;
-use crate::types_structs::{Frag, GnPosition, HapNode, SnpPosition, TraceBackNode, VcfProfile, Options};
+use crate::types_structs::{Frag, GnPosition, HapNode, SnpPosition, TraceBackNode, VcfProfile, Options, FlowUpVec};
 use crate::utils_frags;
 use fxhash::{FxHashMap, FxHashSet};
-use highs::{RowProblem, Sense};
 use rayon::prelude::*;
 use std::sync::Mutex;
 //use osqp::{CscMatrix, Problem, Settings};
@@ -17,7 +16,6 @@ use std::fs::File;
 use std::io::Write;
 use std::mem;
 
-type FlowUpVec = Vec<((usize, usize), (usize, usize), f64)>;
 
 fn update_hap_graph(hap_graph: &mut Vec<Vec<HapNode>>) {
     //    let pseudo_count = 10.;
@@ -100,192 +98,7 @@ fn update_hap_graph(hap_graph: &mut Vec<Vec<HapNode>>) {
     }
 }
 
-pub fn solve_lp_graph(hap_graph: &Vec<Vec<HapNode>>, glopp_out_dir: String) -> FlowUpVec {
-    let mut ae = vec![];
 
-    //LP values
-    let mut pb = RowProblem::default();
-    let mut t = vec![];
-    let mut x = vec![];
-
-    //QP variables
-    // Vector looks like [x,t]
-    let mut num_constraints = 0;
-    let mut a_entry = vec![];
-    let mut p_entry = vec![];
-    let mut l = vec![];
-    let mut u = vec![];
-
-    let mut hap_graph_vec = vec![];
-    for hap_block in hap_graph.iter() {
-        for hap_node in hap_block.iter() {
-            hap_graph_vec.push(hap_node);
-        }
-    }
-
-    let mut edge_to_nodes = vec![];
-    let mut nodes_to_edges = FxHashMap::default();
-
-    for hap_node in hap_graph_vec.iter() {
-        let id1 = hap_node.id;
-        for edge in hap_node.out_edges.iter() {
-            let id2 = hap_graph[hap_node.column + 1][edge.0].id;
-            edge_to_nodes.push((id1, id2));
-            nodes_to_edges.insert((id1, id2), edge_to_nodes.len() - 1);
-            //            ae.push(edge.1 * hap_node.cov());
-            ae.push(edge.1);
-        }
-    }
-
-    for _i in 0..edge_to_nodes.len() {
-        x.push(pb.add_column(0., 0..));
-    }
-    for _i in 0..edge_to_nodes.len() {
-        t.push(pb.add_column(1., 0..));
-    }
-
-    for i in edge_to_nodes.len()..2 * edge_to_nodes.len() {
-        p_entry.push((i, i, 1. / ae[i - edge_to_nodes.len()]));
-    }
-
-    for (column_ind, hap_block) in hap_graph.iter().enumerate() {
-        if column_ind == 0 || column_ind == hap_graph.len() - 1 {
-            continue;
-        }
-        for hap_node in hap_block.iter() {
-            if !hap_node.in_edges.is_empty() && !hap_node.out_edges.is_empty() {
-                let node_id = hap_node.id;
-                let mut in_edge_ids = vec![];
-                let mut out_edge_ids = vec![];
-                for in_edge in hap_node.in_edges.iter() {
-                    let node_id2 = hap_graph[hap_node.column - 1][in_edge.0].id;
-                    let in_edge_id = nodes_to_edges.get(&(node_id2, node_id)).unwrap();
-                    in_edge_ids.push(in_edge_id);
-                }
-                for out_edge in hap_node.out_edges.iter() {
-                    let node_id2 = hap_graph[hap_node.column + 1][out_edge.0].id;
-                    let out_edge_id = nodes_to_edges.get(&(node_id, node_id2)).unwrap();
-                    out_edge_ids.push(out_edge_id);
-                }
-                let mut constraint_row = vec![];
-
-                for in_edge_id in in_edge_ids.iter() {
-                    constraint_row.push((x[**in_edge_id], 1.));
-                }
-                for out_edge_id in out_edge_ids.iter() {
-                    constraint_row.push((x[**out_edge_id], -1.));
-                }
-
-                //                dbg!(&constraint_row, &hap_node.in_edges);
-                pb.add_row(..0, &constraint_row);
-                pb.add_row(0.., &constraint_row);
-
-                //QP variables
-                for in_edge_id in in_edge_ids {
-                    a_entry.push((*in_edge_id, num_constraints, 1.));
-                }
-                for out_edge_id in out_edge_ids {
-                    a_entry.push((*out_edge_id, num_constraints, -1.));
-                }
-                l.push(0.);
-                u.push(0.);
-                num_constraints += 1;
-            }
-        }
-    }
-
-    for i in 0..t.len() {
-        pb.add_row(-1.0 * ae[i].., &[(t[i], 1.), (x[i], -1.)]);
-        pb.add_row(1.0 * ae[i].., &[(t[i], 1.), (x[i], 1.)]);
-        pb.add_row(0.0.., &[(x[i], 1.)]);
-    }
-
-    //QP variables
-    for i in 0..t.len() {
-        a_entry.push((i, num_constraints, 1.));
-        a_entry.push((i + t.len(), num_constraints, -1.));
-        l.push(ae[i]);
-        u.push(ae[i]);
-        num_constraints += 1;
-
-        a_entry.push((i, num_constraints, 1.));
-        l.push(0.);
-        u.push(f64::MAX);
-        num_constraints += 1;
-    }
-
-    //QP SOLVING
-    //    let p = entry_list_to_csc(p_entry, t.len() * 2, t.len() * 2);
-    //    let a = entry_list_to_csc(a_entry, t.len() * 2, num_constraints);
-    //    let mut q = vec![0.; t.len() * 2];
-    //    let lambda = 0.;
-    //    for i in 0..x.len() {
-    //        q[i] = lambda;
-    //    }
-    //
-    //    // Disable verbose output
-    //    let settings = Settings::default().verbose(false);
-    //
-    //    // Create an OSQP problem
-    //    let mut prob = Problem::new(p, &q, a, &l, &u, &settings).expect("failed to setup problem");
-    //
-    //    // Solve problem
-    //    let qp_solution = prob.solve();
-    //    println!("{:?}", qp_solution.x().expect("failed to solve problem"));
-
-    let solved = pb.optimise(Sense::Minimise).solve();
-    let solution = solved.get_solution();
-
-    let mut file = File::create(format!("{}/graph.csv", glopp_out_dir)).expect("Can't create file");
-    for i in 0..edge_to_nodes.len() {
-        if solution.columns()[i] < constants::FLOW_CUTOFF {
-            continue;
-        }
-        writeln!(
-            file,
-            "{},{}-{},{}-{},{}",
-            &solution.columns()[i],
-            hap_graph_vec[edge_to_nodes[i].0].column,
-            edge_to_nodes[i].0,
-            hap_graph_vec[edge_to_nodes[i].1].column,
-            edge_to_nodes[i].1,
-            ae[i]
-        )
-        .unwrap();
-    }
-    drop(file);
-
-    //    let mut file = File::create(format!("{}/qp_graph.csv", glopp_out_dir)).expect("Can't create file");
-    //    for i in 0..edge_to_nodes.len() {
-    //        if qp_solution.x().unwrap()[i] < flow_cutoff {
-    //            continue;
-    //        }
-    //        writeln!(
-    //            file,
-    //            "{},{}-{},{}-{},{}",
-    //            &qp_solution.x().unwrap()[i],
-    //            hap_graph_vec[edge_to_nodes[i].0].column,
-    //            edge_to_nodes[i].0,
-    //            hap_graph_vec[edge_to_nodes[i].1].column,
-    //            edge_to_nodes[i].1,
-    //            ae[i]
-    //        )
-    //        .unwrap();
-    //    }
-    //    drop(file);
-
-    let mut flow_update_vec = vec![];
-    for i in 0..edge_to_nodes.len() {
-        let (node1_id, node2_id) = edge_to_nodes[i];
-        let node1 = hap_graph_vec[node1_id];
-        let node2 = hap_graph_vec[node2_id];
-        let flow = solution.columns()[i];
-        flow_update_vec.push(((node1.column, node1.row), (node2.column, node2.row), flow));
-    }
-
-    log::info!("Linear program finished.");
-    return flow_update_vec;
-}
 
 fn get_local_hap_blocks<'a>(
     all_frags: &'a Vec<Frag>,
@@ -646,7 +459,7 @@ pub fn get_disjoint_paths_rewrite<'a>(
     let mut hap_petgraph = StableGraph::<(usize, usize), f64>::new();
     //Update the graph to include flows.
     for (n1_inf, n2_inf, flow) in flow_update_vec {
-        if flow < constants::FLOW_CUTOFF {
+        if flow < constants::FLOW_CUTOFF_MULT * options.epsilon {
             continue;
         }
         hap_graph[n1_inf.0][n1_inf.1]
