+++ src/lib.rs
@@ -17,3 +17,5 @@ pub mod part_block_manip;
 pub mod constants;
 
 pub mod parse_cmd_line;
+
+pub mod solve_flow;
